+++ crates/alkahest/src/gui/context.rs
@@ -14,7 +14,6 @@ use winit::{event::WindowEvent, window::Window};
 
 use super::sodi::Sodi;
 use crate::{
-    config::APP_DIRS,
     gui::{
         bottom_bar::BottomBar,
         configuration::RenderSettingsPanel,
@@ -30,6 +29,7 @@ use crate::{
         profiler::PuffinProfiler,
         tfx::{TfxErrorViewer, TfxExternEditor},
     },
+    paths,
     resources::AppResources,
     util::image::EguiPngLoader,
 };
@@ -48,7 +48,7 @@ impl GuiContext {
 
         egui.add_image_loader(Arc::new(EguiPngLoader::default()));
 
-        if let Ok(Ok(data)) = std::fs::read_to_string(APP_DIRS.config_dir().join("egui.ron"))
+        if let Ok(Ok(data)) = std::fs::read_to_string(paths::config_dir().join("egui.ron"))
             .map(|s| ron::from_str::<egui::Memory>(&s))
         {
             info!("Loaded egui state from egui.ron");
@@ -159,7 +159,7 @@ impl Drop for GuiContext {
     fn drop(&mut self) {
         match self.egui.memory(ron::to_string) {
             Ok(memory) => {
-                if let Err(e) = std::fs::write(APP_DIRS.config_dir().join("egui.ron"), memory) {
+                if let Err(e) = std::fs::write(paths::config_dir().join("egui.ron"), memory) {
                     error!("Failed to write egui state: {e}");
                 }
             }